requirements.txt
flask
openai
python-dotenv
langchain
langchain_huggingface
langchain_community
sentence-transformers
tiktoken
unstructured
huggingface_hub
g4f
pypdf
pdfplumber
pinecone
curl_cffi
fitz
PyMuPDF
